+++ share/translations/galacteek_fr.ts
@@ -1611,17 +1611,17 @@ Path: {0}, nodes processed: {1}</source>
 <context>
     <name>PyramidMaster</name>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="81"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="82"/>
         <source>Add: dynamic content</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="87"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="88"/>
         <source>Create: image gallery</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="93"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="94"/>
         <source>
         &lt;p&gt;
         Create an image gallery!
@@ -1633,12 +1633,12 @@ Path: {0}, nodes processed: {1}</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="106"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="107"/>
         <source>Add: filesystem synchronizer</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="113"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="114"/>
         <source>
         &lt;p&gt;
         Allows you to &lt;b&gt;synchronize&lt;/b&gt; files (that exist
@@ -1654,12 +1654,12 @@ Path: {0}, nodes processed: {1}</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="131"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="132"/>
         <source>Create: dwebsite (Markdown)</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="145"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="160"/>
         <source>
         &lt;p&gt;
         Create a simple website on the dweb.
@@ -1671,108 +1671,118 @@ Path: {0}, nodes processed: {1}</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="166"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="181"/>
         <source>Open latest content in the pyramid</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="173"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="188"/>
         <source>Pyramid is empty</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="179"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="194"/>
         <source>Pop an item off the pyramid</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="185"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="200"/>
         <source>Force sync</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="191"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="206"/>
         <source>Rewind DAG</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="197"/>
-        <source>Publish</source>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="204"/>
-        <source>Unpublish</source>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="211"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="226"/>
         <source>Register this content as a service in the list of IP services on your DID (Decentralized Identity)</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="218"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="233"/>
         <source>Rewinding the DAG cancels the latest operation/transformation in the DAG&apos;s history</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="225"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="240"/>
         <source>Copy IPNS address to clipboard</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="231"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="246"/>
         <source>Copy IPNS gateway address to clipboard</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="237"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="252"/>
         <source>Add current clipboard item to the pyramid</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="244"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="259"/>
         <source>Generate pyramid&apos;s QR code</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="251"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="266"/>
         <source>Generate index&apos;s QR code</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="258"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="273"/>
         <source>Choose a pyramid to publish this object to</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="265"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="280"/>
         <source>Drop object to a pyramid</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="272"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="287"/>
         <source>Browse image gallery</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="279"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="294"/>
         <source>Browse image gallery (IPNS)</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="286"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="301"/>
         <source>Change image gallery&apos;s title</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="138"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="139"/>
         <source>Create: gem (gemini ipfs capsule)</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../../galacteek/ui/pyramids.py" line="159"/>
+        <location filename="../../galacteek/ui/pyramids.py" line="174"/>
         <source>Access Gemini capsule</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../../galacteek/ui/pyramids.py" line="146"/>
+        <source>Create: http forward service (forward to an existing HTTP service)</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../../galacteek/ui/pyramids.py" line="153"/>
+        <source>Access ipfs+http website</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../../galacteek/ui/pyramids.py" line="212"/>
+        <source>Publish (DID service)</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../../galacteek/ui/pyramids.py" line="219"/>
+        <source>Unpublish (DID service)</source>
         <translation type="unfinished"></translation>
     </message>
 </context>