+++ CHANGELOG.md
@@ -5,6 +5,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.6] - 2022-08-08
+
+### Added
+
+- Functions to reencode PeerIDs into different base encodings
+- Support for the ipfs+http(s):// protocol
+- UI: add actions to create HTTP forward DID services
+
 ## [0.5.5] - 2022-04-23
 
 ### Added